SocioSense2/requirements.txt
streamlit
python-dotenv
google-cloud-aiplatform